requirements.txt
flask
flask-cors
pymongo
python-dotenv
pyjwt
gunicorn
werkzeug
certifi
requests
dnspython
